Implemented type-checking.
There's now a pre-execution analysis pass which detects type
errors through the program. As part of the updates to support
typechecking, I updated the grammar to require type declarations
(I'll probably relax that as I work some type inference into
it), and a total refactor of how the built-in arithmetic operations
are implemented.